+++ src/context/User.js
@@ -6,9 +6,22 @@ import auth from '@react-native-firebase/auth'
 // el valor.
 export const UserContext = createContext('')
 
-const user = "prueba"
-
 export const UserProvider = ({ children }) => {
+
+  const [initializing, setInitializing] = useState(true)
+  const [user, setUser] = useState()
+
+  function onAuthStateChanged(user) {
+    setUser(user)
+    if (initializing) setInitializing(false)
+  }
+
+  useEffect(() => {
+    const subscriber = auth().onAuthStateChanged(onAuthStateChanged)
+    return subscriber
+  }, [])
+
+
   return (
     <UserContext.Provider value={user}>
       {children}
